tests/requirements.txt
django
djangorestframework
opencontainers
